backend/requirements.txt
fastapi
uvicorn
pyodbc
python-dotenv
passlib[bcrypt]
python-jose
